+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 ### Breaking changes
  * Migration to `@animoca/ethereum-contracts-core_library:1.0.0` with `solc:0.6.x` and `@openzeppelin/contracts:3.x`.
+ * `AssetsInventory` derived contracts must now implement internal function `_uri(uint256 id)`.
 
 ### New features
  * Added `Meta20InventoryMock.sol` and related migration.
-
+ * Added internal functions `_mintFungible` and `_mintNonFungible` to `AssetsInventory`.
 
 ## 1.1.2 (23/04/2020)